order-model,order-routes,lineitem-routes-completed

+++ server/db/models/LineItem.js
@@ -10,7 +10,7 @@ const LineItem = db.define('lineitem', {
     allowNull: false
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER
   }
 })
